requirements.txt
certifi==2023.5.7 ; python_version >= "3.7" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.7" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.7" and python_version < "4.0"
dj-static==0.0.6 ; python_version >= "3.7" and python_version < "4.0"
django-extensions==1.9.9 ; python_version >= "3.7" and python_version < "4.0"
django-registration-redux==1.11 ; python_version >= "3.7" and python_version < "4.0"
django==2.0 ; python_version >= "3.7" and python_version < "4.0"
docopt==0.4.0 ; python_version >= "3.7" and python_version < "4.0"
idna==3.4 ; python_version >= "3.7" and python_version < "4.0"
mailchimp==2.0.10 ; python_version >= "3.7" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.7" and python_version < "4.0"
pyjwt==2.7.0 ; python_version >= "3.7" and python_version < "4.0"
python-dotenv==0.18.0 ; python_version >= "3.7" and python_version < "4.0"
python-social-auth==0.2.21 ; python_version >= "3.7" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.7" and python_version < "4.0"
pytz==2015.7 ; python_version >= "3.7" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.7" and python_version < "4.0"
requests==2.30.0 ; python_version >= "3.7" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.7" and python_version < "4.0"
static3==0.7.0 ; python_version >= "3.7" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version == "3.7"
typing==3.7.4.3 ; python_version >= "3.7" and python_version < "4.0"
urllib3==2.0.2 ; python_version >= "3.7" and python_version < "4.0"
